Last components' hierarchy change
