--- src/components/PostPage/PostPage.tsx
@@ -1,7 +1,0 @@
-export const PostPage = () => {
-  return (
-    <>Hello</>
-  );
-}
-
-export default PostPage;
