+++ src/components/SinglePostPage/SinglePostPage.tsx
@@ -1,0 +1,7 @@
+export const SinglePostPage = () => {
+  return (
+    <>Hello</>
+  );
+}
+
+export default SinglePostPage;
